DB schema: decimal precision on PO/SO line + customer creditLimit (Phase 3 F8-partial)
Phase 1 found PO line quantity stored as int32 (no fractional quantities) and
unit price truncating to 4 decimals silently. UoM-aware shops need fractional
quantities — material-by-weight (lb, kg), by-time (hr), by-volume (gal, l).

Change PO line + SO line:
  - Quantity: int -> decimal(18, 4)
  - UnitPrice: -> decimal(18, 4) (industry-standard precision)

Customer.creditLimit confirmed decimal(18, 2) (money precision).

Migration: WU10_DecimalPrecision_POLine_SOLine_CustomerCreditLimit. Existing
int quantities cast cleanly. No data loss.

Scope explicitly excludes Invoice/Payment amounts, BOM quantities, routing
times, ExchangeRate, and other entities — those are deferred to WU-23-F8-broad
per the Phase 3 plan.

Cases addressed: F8-partial (TODO), EDGE-DECIMAL-PRECISION-001, EDGE-DECIMAL-PRECISION-004.

+++ qb-engineer.api/Features/Mrp/ReleasePlannedOrder.cs
@@ -53,7 +53,10 @@ public class ReleasePlannedOrderHandler(AppDbContext db, IClock clock, IBarcodeS
                 PurchaseOrderId = po.Id,
                 PartId = order.PartId,
                 Description = order.Part?.Description ?? "",
-                OrderedQuantity = (int)Math.Ceiling(order.Quantity),
+                // Phase 3 / WU-10 — OrderedQuantity is decimal now; preserve the
+                // Math.Ceiling rounding because MRP planned quantities have always
+                // rounded up (whole-unit purchase increments).
+                OrderedQuantity = Math.Ceiling(order.Quantity),
                 UnitPrice = 0,
                 MrpPlannedOrderId = order.Id,
             };
