+++ qb-engineer.core/Entities/SalesOrderLine.cs
@@ -5,15 +5,19 @@ public class SalesOrderLine : BaseEntity
     public int SalesOrderId { get; set; }
     public int? PartId { get; set; }
     public string Description { get; set; } = string.Empty;
-    public int Quantity { get; set; }
+    // Phase 3 / WU-10 / F8-partial: quantities are decimal, not int. UoM-aware
+    // shops need fractional quantities — material-by-weight (lb, kg), by-time
+    // (hr), by-volume (gal, l). decimal(18, 4). Cases EDGE-DECIMAL-PRECISION-001
+    // / -004.
+    public decimal Quantity { get; set; }
     public decimal UnitPrice { get; set; }
     public int LineNumber { get; set; }
-    public int ShippedQuantity { get; set; }
+    public decimal ShippedQuantity { get; set; }
     public string? Notes { get; set; }
     public int? UomId { get; set; }
 
     public decimal LineTotal => Quantity * UnitPrice;
-    public int RemainingQuantity => Quantity - ShippedQuantity;
+    public decimal RemainingQuantity => Quantity - ShippedQuantity;
     public bool IsFullyShipped => ShippedQuantity >= Quantity;
 
     public SalesOrder SalesOrder { get; set; } = null!;
