+++ qb-engineer.core/Models/SalesOrderLineResponseModel.cs
@@ -1,16 +1,17 @@
 namespace QBEngineer.Core.Models;
 
+// Phase 3 / WU-10 / F8-partial — quantities are decimal (was int).
 public record SalesOrderLineResponseModel(
     int Id,
     int? PartId,
     string? PartNumber,
     string Description,
-    int Quantity,
+    decimal Quantity,
     decimal UnitPrice,
     decimal LineTotal,
     int LineNumber,
-    int ShippedQuantity,
-    int RemainingQuantity,
+    decimal ShippedQuantity,
+    decimal RemainingQuantity,
     bool IsFullyShipped,
     string? Notes,
     List<SalesOrderLineJobModel> Jobs);
